refactor(db): move collection helpers and add docs
